+++ scripts/boostrap/development.ts
@@ -20,3 +20,4 @@ Env.load();
 const environment: Environment = "development";
 Env.define("BABEL_ENV", environment);
 Env.define("NODE_ENV", environment);
+Env.define("UV_THREADPOOL_SIZE", "128");
